preditivo_rais/requirements.txt
numpy
pandas
scikit-learn
joblib
pyyaml
matplotlib
seaborn
